requirements.txt
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
google-genai==0.3.0
gunicorn==21.2.0